+++ ReleaseNotes.md
@@ -10,4 +10,6 @@
   - added example of bad-pixel list  
      (sn2897_badpixels.txt for device with serial numbe 2897
 
-  - added analysis prescaling 
+  - added analysis prescaling
+
+  - added callback mode to speed-up data acquisition - now up to 20 Hz possible